git-svn-id: https://svn.eduke32.com/eduke32@1018 1a8010ca-5511-0410-912e-c29ae57300e0

+++ polymer/eduke32/source/namesdyn.c
@@ -766,7 +766,7 @@ int RESERVEDSLOT10 = 6141;
 int RESERVEDSLOT11 = 6142;
 int RESERVEDSLOT12 = 6143;
 
-int dynamictostatic[MAXTILES];
+short dynamictostatic[MAXTILES];
 
 void processnames(const char *szLabel, int lValue)
 {
